refactor: componentize
