+++ src/core/components/ChatView.module.css
@@ -155,7 +155,7 @@
 
 #input {
   min-height: 50px;
-  max-width: 500px;
+  max-width: 400px;
   width: 100%;
   background-color: var(--colors-bgTertiary);
   border-radius: var(--borderRadius-md);
@@ -182,7 +182,7 @@
 
 #inputContainer {
   width: 100%;
-  max-width: 500px;
+  max-width: 400px;
   display: flex;
   align-items: flex-start;
   position: relative;
@@ -222,7 +222,6 @@
 #controls {
   width: 100%;
   flex-shrink: 0;
-
   display: flex;
   flex-direction: column;
   align-items: center;
